% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_indicator.R
\name{append_indicator}
\alias{append_indicator}
\title{append_indicator}
\usage{
append_indicator(index_frame, df, indicator_name, lastyear = 2009)
}
\description{
This function appends the indicator of interest to the working dataframe
}
\details{
append_indicator function
}
\keyword{append}
