% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{rank}
\usage{
compile(index_frame, lastyear, bysex = FALSE)
}
\description{
This function compiles indicators and generates ranking according to preferences (last year used, sex-dependent or total, and compilation algorithm)
}
\details{
compile function
}
\keyword{clean}
\keyword{ilostat}
