% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helper.R
\name{filter_helper}
\alias{filter_helper}
\title{filter_helper}
\usage{
filter_helper(df, index, bygender = FALSE, lastyear = 2009)
}
\description{
This helper function filters the desired observations from the dataframe of each indicator and appends them to the "index" dataframe. Each dataframe passed through this function must have the following columns: ref_area.label (i.e. country name), sex.label, and obs_value.
}
\details{
filter_helper function
}
\examples{
index <- filter_helper(neet, index, bygender = TRUE, lastyear= 2009)
}
\keyword{filter}
\keyword{helper}
