% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch.R
\name{mismatch}
\alias{mismatch}
\title{mismatch}
\usage{
mismatch(empxeduc, uexeduc)
}
\description{
This function calculates the exucation mismatch, taking raw employment-by-education and unemployment-by-education dataframes as inputs
}
\details{
mismatch function
}
\keyword{clean}
\keyword{ilostat}
