% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_generator.R
\name{rank_generator}
\alias{rank_generator}
\title{rank_generator}
\usage{
rank_generator(dfList, country_list, bygender = FALSE, lastyear = 2009)
}
\description{
This function takes a cleaned list of dataframes for each indicator in the youth labor market index and the list of countries of interest and outputs an index score for each country according to the desired parameters. These parameters are: disaggregation by gender (yes or no), the last year (i.e. the oldest data) to be entered into the index, rescaling method (raw, zscore) and aggregation method across indicators and dimensions (arithmetic mean, geometric mean).
}
\details{
rank_generator function
}
\keyword{bygender}
\keyword{dimensions}
\keyword{generator}
\keyword{index}
\keyword{ranking}
