% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transDim.R
\name{transComp}
\alias{transComp}
\title{transComp}
\usage{
transComp(
  country_list,
  neet,
  relative_unemp,
  empxeduc,
  uexeduc,
  bygender = FALSE,
  lastyear = 2009
)
}
\description{
This function compiles the transition dimension of the YLMILIC by gender or in aggregate from raw dataframes. Takes the gender option and the last acceptable year for data as arguments.
}
\details{
tranComp function
}
\keyword{transition}
